requirements.txt
surya-ocr
torch
datasets
tqdm
pdf2image
Pillow
pytesseract
python-dotenv
numpy
azure-ai-formrecognizer